Fix pre-commit checks
